noificaciones con sweetalert

+++ src/app/pages/employees/details/details.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { NavigationExtras, Router } from '@angular/router';
 import { Employee } from 'src/app/shared/components/models/employee.interface';
+import Swal from 'sweetalert2';
 import { EmployeesService } from '../employees.service';
 
 @Component({
@@ -33,8 +34,27 @@ export class DetailsComponent implements OnInit {
 
   async onGoToDelete(): Promise <void>{
     try {
-      await this.employeesService.onDeleteEmployee(this.employee.id)
-      alert('Deleted')
+      Swal.fire({
+        title: `Seguro que desea eliminar a: ${this.employee.name} ${this.employee.lastName}`,
+        text: "No podras revertir tus acciones",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si,eliminalo!',
+        cancelButtonText:'Cancelar'
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          await this.employeesService.onDeleteEmployee(this.employee.id)
+          .then(()=>{
+            Swal.fire(
+              `${this.employee.name} ${this.employee.lastName}`,
+              'Ha sido eliminado',
+              'success'
+            )
+          })
+        }
+      })
       this.onGoBackToList();
     } catch (error) {
       console.error(error)
